fix(speedDial): Fix intially open bug.
Due to an animation issue, the FAB Speed Dial started in an open
state when using the `md-fling` animation.

Fix by delaying the initial animation until after the component
was fully rendered so that the animation's calculations were correct
and adding some CSS to ensure the first animation fired instantly.

Fixes #6111.