fix(input): Spacing/padding issues.
The spacing and paddin of inputs was incorrect due to recent changes.

Fix by:

 - Altering margins and padding
 - Automatically applying the error spacing only to block-level inupts
 - Adding a new `md-error-spacing` class for inline inputs

We now also allow applying the `md-block` or `md-error-spacing` classes
to the `<form>` element to cascade to all nested input containers.

Update docs/demos accordingly.

BREAKING CHANGE

We no longer automatically add bottom spacing to all instance of the
`<md-input-container>`. It is now applied only if you use the `md-block`
or `md-error-spacing` classes to allow for more compact form inputs.